requirements.txt
# ============================================================================
# Production-Ready RAG System - Complete Dependencies
# ============================================================================

# ----------------------------------------------------------------------------
# Core LangChain Framework
# ----------------------------------------------------------------------------
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6

# ----------------------------------------------------------------------------
# Vector Database and Embeddings
# ----------------------------------------------------------------------------
faiss-cpu==1.7.4
sentence-transformers==2.3.1
huggingface-hub<0.17.0
transformers>=4.21.0
torch>=1.11.0

# ----------------------------------------------------------------------------
# Data Collection
# ----------------------------------------------------------------------------
wikipedia==1.4.0

# ----------------------------------------------------------------------------
# LLM APIs
# ----------------------------------------------------------------------------
# Google Gemini (Primary)
google-generativeai==0.3.2

# OpenAI (Alternative)
openai>=1.30.1

# ----------------------------------------------------------------------------
# Web Framework and UI
# ----------------------------------------------------------------------------
streamlit==1.29.0

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
python-dotenv==1.0.0

# ----------------------------------------------------------------------------
# Scientific Computing
# ----------------------------------------------------------------------------
numpy==1.24.3
pandas==2.0.3

# ----------------------------------------------------------------------------
# Security Layer (Guardrails AI)
# ----------------------------------------------------------------------------
# Guardrails AI for LLM safety and validation
guardrails-ai==0.6.7

# Rate limiting
slowapi==0.1.9

# Audit logging
python-json-logger==2.0.7

# ----------------------------------------------------------------------------
# Observability Layer
# ----------------------------------------------------------------------------
# LLM Observability - Langfuse for tracing
langfuse==2.12.0

# Distributed Tracing - OpenTelemetry
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation>=0.43b0

# Metrics - Prometheus
prometheus-client==0.19.0

# Hallucination Detection - ML libraries
scikit-learn==1.3.2

# ----------------------------------------------------------------------------
# Resilience Layer
# ----------------------------------------------------------------------------
# Retry logic with exponential backoff
tenacity==8.2.3

# ----------------------------------------------------------------------------
# Monitoring Layer
# ----------------------------------------------------------------------------
# Health Check - FastAPI
fastapi==0.109.0
uvicorn==0.27.0

# ----------------------------------------------------------------------------
# Testing Dependencies
# ----------------------------------------------------------------------------
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
coverage==7.3.2

# ----------------------------------------------------------------------------
# Development Tools (Optional)
# ----------------------------------------------------------------------------
# Code formatting
# black==23.12.0

# Linting
# flake8==7.0.0
# pylint==3.0.3

# Type checking
# mypy==1.8.0
